Incorporate requested changes for PR
