--- modules/get_me_some_ice_cream/recognition.py
@@ -1,8 +1,0 @@
-# TODO
-def familiar_customer():
-    return True
-
-
-# TODO
-def get_customer_name():
-    return 'Rafael'
