+++ modules/luigi/states/detection.py
@@ -1,8 +1,6 @@
 import ravestate as rs
 import ravestate_nlp as nlp
 
-from get_me_some_ice_cream.utils import extract_scoops
-
 prop_flavor = rs.Property(name="flavor", allow_read=True, allow_write=True, always_signal_changed=True)
 prop_scoops = rs.Property(name="scoops", allow_read=True, allow_write=True, always_signal_changed=True)
 
@@ -32,3 +30,22 @@ def detect_scoops(ctx: rs.ContextWrapper):
         prop_scoops.write(detected_scoops)
         return rs.Emit()
 
+
+def extract_scoops(prop_ner):
+    for entity, NE in prop_ner:
+        if NE == "CARDINAL" and entity.isdigit():
+            return int(entity)
+        elif NE == "CARDINAL" and isinstance(entity, str):
+            # we assume that no one orders more than 9 scoops of a flavor
+            word2num = {
+                "one": 1,
+                "two": 2,
+                "three": 3,
+                "four": 4,
+                "five": 5,
+                "six": 6,
+                "seven": 7,
+                "eight": 8,
+                "nine": 9,
+            }
+            return word2num[entity]
